rename edgedb client to client-old
